+++ tests/test_v1.py
@@ -45,12 +45,7 @@ def test_metadata_for_variable():
         "columnOrder": 0,
         "datasetName": "Technology Adoption - Isard (1942) and others",
         "nonRedistributable": False,
-        "display": {
-            "name": None,
-            "unit": None,
-            "shortUnit": None,
-            "includeInTable": None,
-        },
+        "display": {},
         "source": {
             "id": 6800,
             "name": "Isard (1942) and others",
@@ -64,28 +59,28 @@ def test_metadata_for_variable():
         "dimensions": {
             "years": {
                 "type": "int",
-                "values": {
-                    "1988": None,
-                    "1989": None,
-                    "1990": None,
-                    "1991": None,
-                    "1992": None,
-                    "1993": None,
-                    "1994": None,
-                    "1995": None,
-                    "1996": None,
-                    "1997": None,
-                    "1998": None,
-                    "1999": None,
-                    "2000": None,
-                    "2001": None,
-                    "2002": None,
-                    "2003": None,
-                },
+                "values": [
+                    {"id": 1988},
+                    {"id": 1989},
+                    {"id": 1990},
+                    {"id": 1991},
+                    {"id": 1992},
+                    {"id": 1993},
+                    {"id": 1994},
+                    {"id": 1995},
+                    {"id": 1996},
+                    {"id": 1997},
+                    {"id": 1998},
+                    {"id": 1999},
+                    {"id": 2000},
+                    {"id": 2001},
+                    {"id": 2002},
+                    {"id": 2003},
+                ],
             },
             "entities": {
                 "type": "int",
-                "values": {"13": {"name": "United States", "code": "USA"}},
+                "values": [{"id": 13, "name": "United States", "code": "USA"}],
             },
         },
     }
